python_draft/requirements.txt
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.88.0
PyPDF2==3.0.1
python-docx==0.8.11
openai==1.3.0
langchain==0.0.350
langchain-community==0.0.9
langchain-text-splitters==0.0.1
langchain-pinecone==0.0.3
sentence-transformers==2.2.2
numpy>=1.21.0,<2.0.0
pandas>=2.0.0,<2.2.0
tiktoken==0.5.1
python-dotenv==1.0.0
requests==2.31.0
pinecone-client==3.0.0
python-pptx==0.6.21
openpyxl==3.1.2
tenacity==8.2.3
aiofiles==23.2.1
asyncio-throttle==1.0.2